requirements.txt
streamlit
google-generativeai
dotenv

pandas
matplotlib 
seaborn